// src/database/types/entry.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * score entry for the score app
 */
export interface Entry {
  /**
   * Entity ID
   */
  id?: number;
  /**
   * The entry's name
   */
  name: string;
  /**
   * a description of the entry
   */
  description?: string;
  /**
   * the position that the entry should show in
   */
  index?: number;
  /**
   * optional team name for the entry
   */
  team?: string;
  /**
   * domain to which this entity belongs
   */
  domain: number;
  /**
   * the ID of the owning User
   */
  owner: number;
}